jormungandr: refacto of helpers
