+++ source/jormungandr/tests/destineo_tests.py
@@ -522,224 +522,6 @@ def sort_destineo_date_timezone_test():
     eq_(response.journeys[1], journey_tc3)
     eq_(response.journeys[2], journey_bss_and_tc)
 
-def bike_duration_walking_test():
-    journey = response_pb2.Journey()
-
-    section = journey.sections.add()
-    section.type = response_pb2.STREET_NETWORK
-    section.street_network.mode = response_pb2.Walking
-    section.duration = 60
-    section = journey.sections.add()
-    section.type = response_pb2.PUBLIC_TRANSPORT
-    section.duration = 80
-    section = journey.sections.add()
-    section.type = response_pb2.STREET_NETWORK
-    section.street_network.mode = response_pb2.Walking
-    section.duration = 60
-
-    eq_(destineo.bike_duration(journey), 0)
-
-def bike_duration_bike_test():
-    journey = response_pb2.Journey()
-
-    section = journey.sections.add()
-    section.type = response_pb2.STREET_NETWORK
-    section.street_network.mode = response_pb2.Bike
-    section.duration = 60
-    section = journey.sections.add()
-    section.type = response_pb2.PUBLIC_TRANSPORT
-    section.duration = 80
-    section = journey.sections.add()
-    section.type = response_pb2.STREET_NETWORK
-    section.street_network.mode = response_pb2.Bike
-    section.duration = 70
-
-    eq_(destineo.bike_duration(journey), 130)
-
-
-def bike_duration_bss_test():
-    journey = response_pb2.Journey()
-
-    section = journey.sections.add()
-    section.type = response_pb2.STREET_NETWORK
-    section.street_network.mode = response_pb2.Walking
-    section.duration = 30
-    section = journey.sections.add()
-    section.type = response_pb2.BSS_RENT
-    section.duration = 10
-    section = journey.sections.add()
-    section.type = response_pb2.STREET_NETWORK
-    section.street_network.mode = response_pb2.Bike
-    section.duration = 120
-    section = journey.sections.add()
-    section.type = response_pb2.BSS_PUT_BACK
-    section.duration = 10
-    section = journey.sections.add()
-    section.type = response_pb2.STREET_NETWORK
-    section.street_network.mode = response_pb2.Walking
-    section.duration = 60
-    section = journey.sections.add()
-    section.type = response_pb2.PUBLIC_TRANSPORT
-    section.duration = 80
-    section = journey.sections.add()
-    section.type = response_pb2.STREET_NETWORK
-    section.street_network.mode = response_pb2.Bike
-    section.duration = 5
-
-    eq_(destineo.bike_duration(journey), 5)
-
-def car_duration_walking_test():
-    journey = response_pb2.Journey()
-
-    section = journey.sections.add()
-    section.type = response_pb2.STREET_NETWORK
-    section.street_network.mode = response_pb2.Walking
-    section.duration = 60
-    section = journey.sections.add()
-    section.type = response_pb2.PUBLIC_TRANSPORT
-    section.duration = 80
-    section = journey.sections.add()
-    section.type = response_pb2.STREET_NETWORK
-    section.street_network.mode = response_pb2.Walking
-    section.duration = 60
-
-    eq_(destineo.car_duration(journey), 0)
-
-def car_duration_car_test():
-    journey = response_pb2.Journey()
-
-    section = journey.sections.add()
-    section.type = response_pb2.LEAVE_PARKING
-    section.duration = 10
-    section = journey.sections.add()
-    section.type = response_pb2.STREET_NETWORK
-    section.street_network.mode = response_pb2.Car
-    section.duration = 60
-    section = journey.sections.add()
-    section.type = response_pb2.PARK
-    section.duration = 10
-    section = journey.sections.add()
-    section.type = response_pb2.PUBLIC_TRANSPORT
-    section.duration = 80
-    section = journey.sections.add()
-    section.type = response_pb2.STREET_NETWORK
-    section.street_network.mode = response_pb2.Walking
-    section.duration = 70
-
-    eq_(destineo.car_duration(journey), 60)
-
-
-def car_duration_bss_test():
-    journey = response_pb2.Journey()
-
-    section = journey.sections.add()
-    section.type = response_pb2.STREET_NETWORK
-    section.street_network.mode = response_pb2.Walking
-    section.duration = 30
-    section = journey.sections.add()
-    section.type = response_pb2.BSS_RENT
-    section.duration = 10
-    section = journey.sections.add()
-    section.type = response_pb2.STREET_NETWORK
-    section.street_network.mode = response_pb2.Bike
-    section.duration = 120
-    section = journey.sections.add()
-    section.type = response_pb2.BSS_PUT_BACK
-    section.duration = 10
-    section = journey.sections.add()
-    section.type = response_pb2.STREET_NETWORK
-    section.street_network.mode = response_pb2.Walking
-    section.duration = 60
-    section = journey.sections.add()
-    section.type = response_pb2.PUBLIC_TRANSPORT
-    section.duration = 80
-    section = journey.sections.add()
-    section.type = response_pb2.LEAVE_PARKING
-    section.duration = 10
-    section = journey.sections.add()
-    section.type = response_pb2.STREET_NETWORK
-    section.street_network.mode = response_pb2.Car
-    section.duration = 180
-    section = journey.sections.add()
-    section.type = response_pb2.PARK
-    section.duration = 10
-
-    eq_(destineo.car_duration(journey), 180)
-
-def tc_duration_walking_test():
-    journey = response_pb2.Journey()
-
-    section = journey.sections.add()
-    section.type = response_pb2.STREET_NETWORK
-    section.street_network.mode = response_pb2.Walking
-    section.duration = 60
-    section = journey.sections.add()
-    section.type = response_pb2.PUBLIC_TRANSPORT
-    section.duration = 80
-    section = journey.sections.add()
-    section.type = response_pb2.STREET_NETWORK
-    section.street_network.mode = response_pb2.Walking
-    section.duration = 60
-
-    eq_(destineo.tc_duration(journey), 80)
-
-def tc_duration_bike_test():
-    journey = response_pb2.Journey()
-
-    section = journey.sections.add()
-    section.type = response_pb2.STREET_NETWORK
-    section.street_network.mode = response_pb2.Bike
-    section.duration = 60
-    section = journey.sections.add()
-    section.type = response_pb2.PUBLIC_TRANSPORT
-    section.duration = 80
-    section = journey.sections.add()
-    section.type = response_pb2.TRANSFER
-    section.duration = 20
-    section = journey.sections.add()
-    section.type = response_pb2.WAITING
-    section.duration = 25
-    section = journey.sections.add()
-    section.type = response_pb2.PUBLIC_TRANSPORT
-    section.duration = 40
-    section = journey.sections.add()
-    section.type = response_pb2.STREET_NETWORK
-    section.street_network.mode = response_pb2.Walking
-    section.duration = 70
-
-    eq_(destineo.tc_duration(journey), 120)
-
-
-def tc_duration_bss_test():
-    journey = response_pb2.Journey()
-
-    section = journey.sections.add()
-    section.type = response_pb2.STREET_NETWORK
-    section.street_network.mode = response_pb2.Walking
-    section.duration = 30
-    section = journey.sections.add()
-    section.type = response_pb2.BSS_RENT
-    section.duration = 10
-    section = journey.sections.add()
-    section.type = response_pb2.STREET_NETWORK
-    section.street_network.mode = response_pb2.Bike
-    section.duration = 120
-    section = journey.sections.add()
-    section.type = response_pb2.BSS_PUT_BACK
-    section.duration = 10
-    section = journey.sections.add()
-    section.type = response_pb2.STREET_NETWORK
-    section.street_network.mode = response_pb2.Walking
-    section.duration = 60
-    section = journey.sections.add()
-    section.type = response_pb2.PUBLIC_TRANSPORT
-    section.duration = 80
-    section = journey.sections.add()
-    section.type = response_pb2.STREET_NETWORK
-    section.street_network.mode = response_pb2.Bike
-    section.duration = 5
-
-    eq_(destineo.tc_duration(journey), 80)
 
 class Instance(object):
     def __init__(self):
@@ -1189,203 +971,6 @@ def get_bss_car_journey():
     section = journey.sections.add()
 
     return journey
-
-
-#==============================================================================
-# has_bss_first_and_walking_last test
-#==============================================================================
-def has_bss_first_and_walking_last__walking_walking_test():
-    assert not destineo.has_bss_first_and_walking_last(get_walking_walking_journey())
-
-def has_bss_first_and_walking_last__walking_bike_test():
-    assert not destineo.has_bss_first_and_walking_last(get_walking_bike_journey())
-
-def has_bss_first_and_walking_last__walking_bss_test():
-    assert not destineo.has_bss_first_and_walking_last(get_walking_bss_journey())
-
-def has_bss_first_and_walking_last__walking_car_test():
-    assert not destineo.has_bss_first_and_walking_last(get_walking_car_journey())
-
-def has_bss_first_and_walking_last__bss_walking_test():
-    assert destineo.has_bss_first_and_walking_last(get_bss_walking_journey())
-
-def has_bss_first_and_walking_last__bss_bike_test():
-    assert not destineo.has_bss_first_and_walking_last(get_bss_bike_journey())
-
-def has_bss_first_and_walking_last__bss_bss_test():
-    assert not destineo.has_bss_first_and_walking_last(get_bss_bss_journey())
-
-def has_bss_first_and_walking_last__bss_car_test():
-    assert not destineo.has_bss_first_and_walking_last(get_bss_car_journey())
-
-def has_bss_first_and_walking_last__bike_walking_test():
-    assert not destineo.has_bss_first_and_walking_last(get_bike_walking_journey())
-
-def has_bss_first_and_walking_last__bike_bike_test():
-    assert not destineo.has_bss_first_and_walking_last(get_bike_bike_journey())
-
-def has_bss_first_and_walking_last__bike_bss_test():
-    assert not destineo.has_bss_first_and_walking_last(get_bike_bss_journey())
-
-def has_bss_first_and_walking_last__bike_car_test():
-    assert not destineo.has_bss_first_and_walking_last(get_bike_car_journey())
-
-
-#==============================================================================
-# has_walking_first_and_bss_last test
-#==============================================================================
-def has_walking_first_and_bss_last__walking_walking_test():
-    assert not destineo.has_walking_first_and_bss_last(get_walking_walking_journey())
-
-def has_walking_first_and_bss_last__walking_bike_test():
-    assert not destineo.has_walking_first_and_bss_last(get_walking_bike_journey())
-
-def has_walking_first_and_bss_last__walking_bss_test():
-    assert destineo.has_walking_first_and_bss_last(get_walking_bss_journey())
-
-def has_walking_first_and_bss_last__walking_car_test():
-    assert not destineo.has_walking_first_and_bss_last(get_walking_car_journey())
-
-def has_walking_first_and_bss_last__bss_walking_test():
-    assert not destineo.has_walking_first_and_bss_last(get_bss_walking_journey())
-
-def has_walking_first_and_bss_last__bss_bike_test():
-    assert not destineo.has_walking_first_and_bss_last(get_bss_bike_journey())
-
-def has_walking_first_and_bss_last__bss_bss_test():
-    assert not destineo.has_walking_first_and_bss_last(get_bss_bss_journey())
-
-def has_walking_first_and_bss_last__bss_car_test():
-    assert not destineo.has_walking_first_and_bss_last(get_bss_car_journey())
-
-def has_walking_first_and_bss_last__bike_walking_test():
-    assert not destineo.has_walking_first_and_bss_last(get_bike_walking_journey())
-
-def has_walking_first_and_bss_last__bike_bike_test():
-    assert not destineo.has_walking_first_and_bss_last(get_bike_bike_journey())
-
-def has_walking_first_and_bss_last__bike_bss_test():
-    assert not destineo.has_walking_first_and_bss_last(get_bike_bss_journey())
-
-def has_walking_first_and_bss_last__bike_car_test():
-    assert not destineo.has_walking_first_and_bss_last(get_bike_car_journey())
-
-#==============================================================================
-# has_bss_first_and_bss_last test
-#==============================================================================
-def has_bss_first_and_bss_last__walking_walking_test():
-    assert not destineo.has_bss_first_and_bss_last(get_walking_walking_journey())
-
-def has_bss_first_and_bss_last__walking_bike_test():
-    assert not destineo.has_bss_first_and_bss_last(get_walking_bike_journey())
-
-def has_bss_first_and_bss_last__walking_bss_test():
-    assert not destineo.has_bss_first_and_bss_last(get_walking_bss_journey())
-
-def has_bss_first_and_bss_last__walking_car_test():
-    assert not destineo.has_bss_first_and_bss_last(get_walking_car_journey())
-
-def has_bss_first_and_bss_last__bss_walking_test():
-    assert not destineo.has_bss_first_and_bss_last(get_bss_walking_journey())
-
-def has_bss_first_and_bss_last__bss_bike_test():
-    assert not destineo.has_bss_first_and_bss_last(get_bss_bike_journey())
-
-def has_bss_first_and_bss_last__bss_bss_test():
-    assert destineo.has_bss_first_and_bss_last(get_bss_bss_journey())
-
-def has_bss_first_and_bss_last__bss_car_test():
-    assert not destineo.has_bss_first_and_bss_last(get_bss_car_journey())
-
-def has_bss_first_and_bss_last__bike_walking_test():
-    assert not destineo.has_bss_first_and_bss_last(get_bike_walking_journey())
-
-def has_bss_first_and_bss_last__bike_bike_test():
-    assert not destineo.has_bss_first_and_bss_last(get_bike_bike_journey())
-
-def has_bss_first_and_bss_last__bike_bss_test():
-    assert not destineo.has_bss_first_and_bss_last(get_bike_bss_journey())
-
-def has_bss_first_and_bss_last__bike_car_test():
-    assert not destineo.has_bss_first_and_bss_last(get_bike_car_journey())
-
-#==============================================================================
-# has_bike_first_and_walking_last test
-#==============================================================================
-def has_bike_first_and_walking_last__walking_walking_test():
-    assert not destineo.has_bike_first_and_walking_last(get_walking_walking_journey())
-
-def has_bike_first_and_walking_last__walking_bike_test():
-    assert not destineo.has_bike_first_and_walking_last(get_walking_bike_journey())
-
-def has_bike_first_and_walking_last__walking_bss_test():
-    assert not destineo.has_bike_first_and_walking_last(get_walking_bss_journey())
-
-def has_bike_first_and_walking_last__walking_car_test():
-    assert not destineo.has_bike_first_and_walking_last(get_walking_car_journey())
-
-def has_bike_first_and_walking_last__bss_walking_test():
-    assert not destineo.has_bike_first_and_walking_last(get_bss_walking_journey())
-
-def has_bike_first_and_walking_last__bss_bike_test():
-    assert not destineo.has_bike_first_and_walking_last(get_bss_bike_journey())
-
-def has_bike_first_and_walking_last__bss_bss_test():
-    assert not destineo.has_bike_first_and_walking_last(get_bss_bss_journey())
-
-def has_bike_first_and_walking_last__bss_car_test():
-    assert not destineo.has_bike_first_and_walking_last(get_bss_car_journey())
-
-def has_bike_first_and_walking_last__bike_walking_test():
-    assert destineo.has_bike_first_and_walking_last(get_bike_walking_journey())
-
-def has_bike_first_and_walking_last__bike_bike_test():
-    assert not destineo.has_bike_first_and_walking_last(get_bike_bike_journey())
-
-def has_bike_first_and_walking_last__bike_bss_test():
-    assert not destineo.has_bike_first_and_walking_last(get_bike_bss_journey())
-
-def has_bike_first_and_walking_last__bike_car_test():
-    assert not destineo.has_bike_first_and_walking_last(get_bike_car_journey())
-
-#==============================================================================
-# has_bike_first_and_bss_last test
-#==============================================================================
-def has_bike_first_and_bss_last__walking_walking_test():
-    assert not destineo.has_bike_first_and_bss_last(get_walking_walking_journey())
-
-def has_bike_first_and_bss_last__walking_bike_test():
-    assert not destineo.has_bike_first_and_bss_last(get_walking_bike_journey())
-
-def has_bike_first_and_bss_last__walking_bss_test():
-    assert not destineo.has_bike_first_and_bss_last(get_walking_bss_journey())
-
-def has_bike_first_and_bss_last__walking_car_test():
-    assert not destineo.has_bike_first_and_bss_last(get_walking_car_journey())
-
-def has_bike_first_and_bss_last__bss_walking_test():
-    assert not destineo.has_bike_first_and_bss_last(get_bss_walking_journey())
-
-def has_bike_first_and_bss_last__bss_bike_test():
-    assert not destineo.has_bike_first_and_bss_last(get_bss_bike_journey())
-
-def has_bike_first_and_bss_last__bss_bss_test():
-    assert not destineo.has_bike_first_and_bss_last(get_bss_bss_journey())
-
-def has_bike_first_and_bss_last__bss_car_test():
-    assert not destineo.has_bike_first_and_bss_last(get_bss_car_journey())
-
-def has_bike_first_and_bss_last__bike_walking_test():
-    assert not destineo.has_bike_first_and_bss_last(get_bike_walking_journey())
-
-def has_bike_first_and_bss_last__bike_bike_test():
-    assert not destineo.has_bike_first_and_bss_last(get_bike_bike_journey())
-
-def has_bike_first_and_bss_last__bike_bss_test():
-    assert destineo.has_bike_first_and_bss_last(get_bike_bss_journey())
-
-def has_bike_first_and_bss_last__bike_car_test():
-    assert not destineo.has_bike_first_and_bss_last(get_bike_car_journey())
 
 def choose_best_alternatives_simple_test():
     journeys = [get_bss_bss_journey(), get_bike_car_journey(), get_bss_walking_journey()]
